@@ -1,5 +1,7 @@
-<script>
+<script lang="ts">
 	import '../global.css';
+
+	let { children } = $props();
 </script>
 
-<slot />
+{@render children()}
